test: add test files/resources
Created a cpp file for Vector tile test.
Added the test function in the .h file for the unit tests.
Added a pbf file for the world map at zoom level 0 to be used for the unit tests.
updated the CMakeLists file to include the protobuf file in the test resources.

Signed-off-by: Eimen Oueslati <aymanoueslati22@gmail.com>