feat/fix: update networking and tileLoading classes to handle errors more gracefully
std::option was previously used by the network controller to store a potential result. This has been changed to the custom httpResponse struct type that stores data with an error code (enum). Errors are printed to qWarning(). Loading and completion of steps is printed to qDebug() to show what parts of the program complete (or not).

Updates: #47