fix: fix app crash bug
Fixed a bug where the app was crashing if we try to load tiles at zoom levels beyond 12.
The cause of the bug was that some properties from the style sheet layers were coded as QVariant and
were treated as QList when the underlying type was not a list, which caused the conversion from QVariant
to QList to produce an empty list, and the app would crash when we try to access the elements of the empty list.

Signed-off-by: Eimen Oueslati <aymanoueslati22@gmail.com>