test/refactor(rendering): refactor a normalization lambda into external function and unit test it
The normalization function sets a value in the range [0, 1]. Some of the maths has a divisor that can approach 0. Mathematically, the result would then approach infinity. We set the result to 0 instead to force the value to be in the [0, 1] range.

The unit tests check for both a positive and negative case. The tests pass.